Move `DisplayAccount` to `: feature:navigation:drawer` domain
